requirements.txt
pandas
jupyter
scipy
nltk
xgboost
torchtext
seaborn
scikit-learn
tensorboard
textdistance
